+++ build_tools/aws-sdk-code-generator/lib/aws-sdk-code-generator/service.rb
@@ -20,6 +20,7 @@ module AwsSdkCodeGenerator
     # @option options [Hash<gem,version>] :gem_dependencies ({})
     # @option options [Hash] :add_plugins ({})
     # @option options [Hash] :remove_plugins ([])
+    # @option options [Boolean] :deprecated (false)
     def initialize(options)
       @name = options.fetch(:name)
       @identifier = name.downcase
@@ -57,6 +58,7 @@ module AwsSdkCodeGenerator
       @require_endpoint_discovery = api.fetch('operations', []).any? do |_, o|
         o['endpointdiscovery'] && o['endpointdiscovery']['required']
       end
+      @deprecated = options[:deprecated] || false
     end
 
     # @return [String] The service name, e.g. "S3"
@@ -145,6 +147,11 @@ module AwsSdkCodeGenerator
 
     # @return [Boolean] true if any operation requires endpoint_discovery
     attr_reader :require_endpoint_discovery
+
+    # @return [Boolean] true if the service is deprecated
+    def deprecated?
+      @deprecated
+    end
 
     # @api private
     def inspect